feat: edit sql schema and add generate.go

+++ app/logger/job/internal/data/ent/sensor/where.go
@@ -9,28 +9,28 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int64) predicate.Sensor {
+func ID(id int) predicate.Sensor {
 	return predicate.Sensor(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldID), id))
 	})
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int64) predicate.Sensor {
+func IDEQ(id int) predicate.Sensor {
 	return predicate.Sensor(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldID), id))
 	})
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int64) predicate.Sensor {
+func IDNEQ(id int) predicate.Sensor {
 	return predicate.Sensor(func(s *sql.Selector) {
 		s.Where(sql.NEQ(s.C(FieldID), id))
 	})
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int64) predicate.Sensor {
+func IDIn(ids ...int) predicate.Sensor {
 	return predicate.Sensor(func(s *sql.Selector) {
 		// if not arguments were provided, append the FALSE constants,
 		// since we can't apply "IN ()". This will make this predicate falsy.
@@ -47,7 +47,7 @@ func IDIn(ids ...int64) predicate.Sensor {
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int64) predicate.Sensor {
+func IDNotIn(ids ...int) predicate.Sensor {
 	return predicate.Sensor(func(s *sql.Selector) {
 		// if not arguments were provided, append the FALSE constants,
 		// since we can't apply "IN ()". This will make this predicate falsy.
@@ -64,28 +64,28 @@ func IDNotIn(ids ...int64) predicate.Sensor {
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int64) predicate.Sensor {
+func IDGT(id int) predicate.Sensor {
 	return predicate.Sensor(func(s *sql.Selector) {
 		s.Where(sql.GT(s.C(FieldID), id))
 	})
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int64) predicate.Sensor {
+func IDGTE(id int) predicate.Sensor {
 	return predicate.Sensor(func(s *sql.Selector) {
 		s.Where(sql.GTE(s.C(FieldID), id))
 	})
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int64) predicate.Sensor {
+func IDLT(id int) predicate.Sensor {
 	return predicate.Sensor(func(s *sql.Selector) {
 		s.Where(sql.LT(s.C(FieldID), id))
 	})
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int64) predicate.Sensor {
+func IDLTE(id int) predicate.Sensor {
 	return predicate.Sensor(func(s *sql.Selector) {
 		s.Where(sql.LTE(s.C(FieldID), id))
 	})
